Moved crate version info into tox_node context
